// src/models/IAccountAddressGeneratorState.ts
// Copyright 2020 IOTA Stiftung
// SPDX-License-Identifier: Apache-2.0

/**
 * Definition of address generator state.
 */
export interface IAccountAddressGeneratorState {
    /**
     * The account index.
     */
    accountIndex: number;

    /**
     * Is this an internal address.
     */
    isInternal: boolean;

    /**
     * The address index.
     */
    addressIndex: number;
}